pyproject.toml
[project]
name = "DockBot"
version = "0.1.0"
description = ""
authors = ["Oisín Mc Laughlin"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
]
